Renaming bot.py -> autobot.py, updating docs
